# Slight improvement in performance